+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* Fixed bug where the arrow navigation didn't work for the dropdown component in
+  combination with a search query.
 * Fixed bug where the page refreshes if you press enter in an input in the row modal.
 
 ## Released (2020-12-01)